Extracted evidencecheckable from Application
This is just an example of how we can do things: instead of having it
all stuffed into the Application model, tease it out.

The new nested structure directory 'query', that is open for debate,
maybe it should be called 'queries'?
