+++ app/models/application.rb
@@ -11,13 +11,6 @@ class Application < ActiveRecord::Base # rubocop:disable ClassLength
 
   validates :reference, presence: true, uniqueness: true
 
-  scope :evidencecheckable, lambda {
-    where(benefits: false,
-          application_type: 'income',
-          emergency_reason: nil,
-          application_outcome: %w[part full])
-  }
-
   scope :waiting_for_evidence, lambda {
     includes(:evidence_check).
       references(:evidence_check).
