+++ spec/models/application_spec.rb
@@ -199,23 +199,6 @@ RSpec.describe Application, type: :model do
       end
     end
 
-    describe '.evidencecheckable' do
-      subject { described_class.evidencecheckable }
-
-      let!(:application_1) { create :application_part_remission }
-      let!(:application_2) { create :application_full_remission }
-      let!(:application_3) { create :application_no_remission }
-      let!(:emergency_application) { create :application_full_remission, emergency_reason: 'REASON' }
-
-      it 'includes only part and full remission applications' do
-        is_expected.to match_array([application_1, application_2])
-      end
-
-      it 'does not include emergency applications' do
-        is_expected.not_to include emergency_application
-      end
-    end
-
     describe '.waiting_for_evidence', focus: true do
       let!(:application1) { create :application }
       let!(:application2) { create :application }
